Update audit report and fix additional kernel bugs
This change marks bugs 3-121 as Fixed in audit_report.md, as they were verified to be addressed in the codebase.
It also fixes several new bugs (122-128) identified during the audit update:
- Bug 122 (Locking): Check mutex_lock return value.
- Bug 123 (Null Check): Null check in select_thread.
- Bug 124 (Allocation Size): Sanity check in KMessage::ReceiveFrom.
- Bug 125 (Integer Overflow): Safe arithmetic in ELF parsing.
- Bug 126 (Permissions): Add creator ID checks in XsiMessageQueue.
- Bug 127 (File Cache Overflows): Bounds/overflow checks in file_cache.cpp.
- Bug 128 (Zombie Node): Fix device_node cleanup on error.

Additionally, this addresses a nitpick in `file_cache_read` by replacing an unsafe cast to `off_t` with a safe `uint64` comparison to prevent potential issues with large file sizes on 64-bit systems.

+++ src/system/kernel/cache/file_cache.cpp
@@ -1469,7 +1469,7 @@ file_cache_read(void* _cacheRef, void* cookie, off_t offset, void* buffer,
 		*_size = 0;
 		return B_OK;
 	}
-	if ((off_t)*_size > fileSize - offset)
+	if ((uint64)*_size > (uint64)(fileSize - offset))
 		*_size = fileSize - offset;
 
 	if (ref->disabled_count > 0) {
